moving name casing features to utils
